hooks moved to new folder, generics added/updated
