+++ apps/frontend/src/api/hooks/tasks/useDeleteTask.ts
@@ -1,9 +1,12 @@
-import { axiosInstance } from '../utilities/axiosInstance';
-import { apiConfig } from '../config/apiRoutes';
+import { axiosInstance } from '../../utilities/axiosInstance';
+import { apiConfig } from '../../config/apiRoutes';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { Task } from '../../types/task';
 
 const deleteTask = async (_id: string | undefined) => {
-  await axiosInstance.delete(`${apiConfig.updateTaskStatus.endpoint}${_id}`);
+  await axiosInstance.delete<Task>(
+    `${apiConfig.updateTaskStatus.endpoint}${_id}`,
+  );
 };
 
 export const useDeleteTask = () => {
