+++ apps/frontend/src/api/hooks/tasks/useGetExampleTaskByContent.ts
@@ -1,11 +1,11 @@
-import { axiosInstance } from '../utilities/axiosInstance';
-import { apiConfig } from '../config/apiRoutes';
-import { TasksResponse } from '../types/task';
 import { useQuery } from '@tanstack/react-query';
+import { axiosInstance } from '../../utilities/axiosInstance';
+import { apiConfig } from '../../config/apiRoutes';
+import { Tasks } from '../../types/task';
 import { useGetExampleTasks } from './useGetExampleTasks';
 
 const getExampleTaskByContent = async (query: string | null) => {
-  const response = await axiosInstance.get<TasksResponse>(
+  const response = await axiosInstance.get<Tasks>(
     `${apiConfig.getExampleTaskByContent.endpoint}` + query,
   );
   return response.data;
