+++ apps/frontend/src/api/hooks/tasks/useMarkTaskAsDone.ts
@@ -1,11 +1,15 @@
-import { axiosInstance } from '../utilities/axiosInstance';
-import { apiConfig } from '../config/apiRoutes';
+import { axiosInstance } from '../../utilities/axiosInstance';
+import { apiConfig } from '../../config/apiRoutes';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { Tasks } from '../../types/task';
 
 const markTaskAsDone = async (_id?: string) => {
-  await axiosInstance.patch(`${apiConfig.updateTaskStatus.endpoint}${_id}`, {
-    done: true,
-  });
+  await axiosInstance.patch<Tasks>(
+    `${apiConfig.updateTaskStatus.endpoint}${_id}`,
+    {
+      done: true,
+    },
+  );
 };
 
 export const useMarkTaskAsDone = () => {
