+++ apps/frontend/src/api/hooks/tasks/usePostTask.ts
@@ -1,9 +1,10 @@
-import { axiosInstance } from '../utilities/axiosInstance';
-import { apiConfig } from '../config/apiRoutes';
+import { axiosInstance } from '../../utilities/axiosInstance';
+import { apiConfig } from '../../config/apiRoutes';
 import { useMutation, useQueryClient } from '@tanstack/react-query';
+import { Task } from '../../types/task';
 
 const postTask = async (task: { content: string; done: boolean }) => {
-  await axiosInstance.post(`${apiConfig.postTask.endpoint}`, task);
+  await axiosInstance.post<Task>(`${apiConfig.postTask.endpoint}`, task);
 };
 
 export const usePostTask = () => {
