+++ apps/frontend/src/to-do-list/TasksPage/TasksList/tasksList.tsx
@@ -6,10 +6,10 @@ import { Content } from '../../../components/Content/content';
 import searchQueryParamName from '../../../utils/searchQueryParamName';
 import classNames from 'classnames';
 import { useLocation } from 'react-router-dom';
-import { useGetTasks } from '../../../api/hooks/useGetTasks';
-import { useDeleteTask } from '../../../api/hooks/useDeleteTask';
-import { useMarkTaskAsDone } from '../../../api/hooks/useMarkTaskAsDone';
-import { useMarkTaskAsUndone } from '../../../api/hooks/useMarkTaskAsUndone';
+import { useGetTasks } from '../../../api/hooks/tasks/useGetTasks';
+import { useDeleteTask } from '../../../api/hooks/tasks/useDeleteTask';
+import { useMarkTaskAsDone } from '../../../api/hooks/tasks/useMarkTaskAsDone';
+import { useMarkTaskAsUndone } from '../../../api/hooks/tasks/useMarkTaskAsUndone';
 import DeleteIcon from '../../../components/Buttons/ButtonIcons/deleteButton.svg?react';
 import useTasksStore from '../../../utils/taskStore';
 // import { useEffect } from 'react';
